Support an alternate naming format for PRs
Adding a fallback PR naming format. Specifically because it's common to use the card URI as a branch name, and when there are multiple commits in a PR, GitHub defaults to naming the PR using the name of the branch (with some edits like replacing dashes with spaces), so it'd be nice to catch those if someone forgets to update to the other expected format.